LibWeb: Split FilterValueListStyleValue out of StyleValue.{h,cpp}
