+++ Userland/Libraries/LibWeb/CSS/StyleValue.h
@@ -145,58 +145,6 @@ struct EdgeRect {
     bool operator==(EdgeRect const&) const = default;
 };
 
-namespace Filter {
-
-struct Blur {
-    Optional<Length> radius {};
-    float resolved_radius(Layout::Node const&) const;
-    bool operator==(Blur const&) const = default;
-};
-
-struct DropShadow {
-    Length offset_x;
-    Length offset_y;
-    Optional<Length> radius {};
-    Optional<Color> color {};
-    struct Resolved {
-        float offset_x;
-        float offset_y;
-        float radius;
-        Color color;
-    };
-    Resolved resolved(Layout::Node const&) const;
-    bool operator==(DropShadow const&) const = default;
-};
-
-struct HueRotate {
-    struct Zero {
-        bool operator==(Zero const&) const = default;
-    };
-    using AngleOrZero = Variant<Angle, Zero>;
-    Optional<AngleOrZero> angle {};
-    float angle_degrees() const;
-    bool operator==(HueRotate const&) const = default;
-};
-
-struct Color {
-    enum class Operation {
-        Brightness,
-        Contrast,
-        Grayscale,
-        Invert,
-        Opacity,
-        Saturate,
-        Sepia
-    } operation;
-    Optional<NumberPercentage> amount {};
-    float resolved_amount() const;
-    bool operator==(Color const&) const = default;
-};
-
-};
-
-using FilterFunction = Variant<Filter::Blur, Filter::DropShadow, Filter::HueRotate, Filter::Color>;
-
 // FIXME: Find a better place for this helper.
 inline Gfx::Painter::ScalingMode to_gfx_scaling_mode(CSS::ImageRendering css_value)
 {
@@ -711,34 +659,6 @@ private:
 
     ResolvedType m_resolved_type;
     NonnullOwnPtr<CalcSum> m_expression;
-};
-
-class FilterValueListStyleValue final : public StyleValueWithDefaultOperators<FilterValueListStyleValue> {
-public:
-    static ValueComparingNonnullRefPtr<FilterValueListStyleValue> create(
-        Vector<FilterFunction> filter_value_list)
-    {
-        VERIFY(filter_value_list.size() >= 1);
-        return adopt_ref(*new FilterValueListStyleValue(move(filter_value_list)));
-    }
-
-    Vector<FilterFunction> const& filter_value_list() const { return m_filter_value_list; }
-
-    virtual ErrorOr<String> to_string() const override;
-
-    virtual ~FilterValueListStyleValue() override = default;
-
-    bool properties_equal(FilterValueListStyleValue const& other) const { return m_filter_value_list == other.m_filter_value_list; };
-
-private:
-    FilterValueListStyleValue(Vector<FilterFunction> filter_value_list)
-        : StyleValueWithDefaultOperators(Type::FilterValueList)
-        , m_filter_value_list(move(filter_value_list))
-    {
-    }
-
-    // FIXME: No support for SVG filters yet
-    Vector<FilterFunction> m_filter_value_list;
 };
 
 class FlexStyleValue final : public StyleValueWithDefaultOperators<FlexStyleValue> {
